backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
SQLAlchemy
psycopg2-binary
databases
pydantic
pgvector